Creación de la vista de turnos por estudiante con descargas pdf

+++ src/main/java/com/app/springapp/interfacesServicios/IServicioEstudiante.java
@@ -8,5 +8,5 @@ public interface IServicioEstudiante {
     public void guardarEstudiante(Estudiante estudiante);
     public void eliminarEstudiante(int id);
     public List<Estudiante> obtenerTodos();
-    public Estudiante buscarPorId(int id);
+    public Estudiante buscarPorId(Long id);
 }
